# internal/api/api.swagger.yaml
openapi: 3.0.3
info:
  title: Logiflow API
  description: API для логистической информационной системы
  version: 0.2.0
  contact:
    name: anxi0uz
servers:
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        status:
          type: integer
          nullable: false
        data:
          type: object
          nullable: true
        success:
          type: boolean
          nullable: false
        requestID:
          type: string
          nullable: false
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        requestID:
          type: string

    # ─── Auth ─────────────────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...
    OrderCreate:
      type: object
      required: [destinationAddress]
      properties:
        originWarehouseId:
          type: string
          format: uuid
          nullable: true
        originAddress:
          type: string
          nullable: true
        destinationAddress:
          type: string
        cargoDescription:
          type: string
          nullable: true
        weightKg:
          type: number
          nullable: true
        volumeM3:
          type: number
          nullable: true
        destinationWarehouseId:
          type: string
          format: uuid
          nullable: true
        pickupFrom:
          type: string
          format: date-time
          nullable: true
        pickupTo:
          type: string
          format: date-time
          nullable: true
...
    OrderCancel:
      type: object
      properties:
        reasonCode:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true

    # ─── Notification ─────────────────────────────────────────────────────────
...
paths:
...
  /auth/register:
    post:
      operationId: authRegister
      summary: Регистрация клиента
      description: Публичный эндпоинт — создаёт аккаунт с ролью client. Для создания менеджеров и водителей используйте POST /managers и POST /drivers (только admin).
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Пользователь создан, токены выданы
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email уже занят
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: authLogin
      summary: Авторизация
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Неверный email или пароль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: authRefresh
      summary: Обновление access-токена
      tags: [Auth]
      security: []
      responses:
        "200":
          description: Токены обновлены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Недействительный refresh-токен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/orders:
    get:
      operationId: listOrders
      summary: Список заявок
      tags: [Orders]
      description: >
        Клиент видит только свои заявки, менеджер — заявки своего склада,
        водитель — назначенные на него, администратор — все.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, ready_for_dispatch, assigned, in_transit, arrived, completed, cancelled]
        - name: driverId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Список заявок
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

    post:
      operationId: createOrder
      summary: Создать заявку
      tags: [Orders]
      description: >
        Создаёт черновик заявки, геокодирует адреса через Nominatim и строит маршрут через OSRM.
        Координаты сохраняются в таблице routes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreate"
      responses:
        "201":
          description: Заявка создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Некорректные данные или сервис маршрутизации недоступен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: getOrder
      summary: Получить заявку
      tags: [Orders]
      responses:
        "200":
          description: Данные заявки
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "404":
          description: Не найдено
  /api/v1/orders/{id}/submit:
    post:
      operationId: submitOrder
      summary: Подготовить заказ к диспетчеризации
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Заказ готов к диспетчеризации }
        "409": { description: Недопустимый переход }
  /api/v1/orders/{id}/cancel:
    post:
      operationId: cancelOrder
      summary: Отменить заказ
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderCancel" }
      responses:
        "200": { description: Заказ отменён }
        "409": { description: Заказ нельзя отменить }
...
  /orders/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    patch:
      operationId: updateOrderStatus
      summary: Устаревшее универсальное обновление статуса
      tags: [Orders]
      deprecated: true
      description: Всегда возвращает 410. Используйте явные команды `/api/v1/orders/...` и `/api/v1/assignments/...`.
      responses:
        "410":
          description: Универсальное изменение статуса удалено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

  # ─── Routes ─────────────────────────────────────────────────────────────────
...
  /reports/orders:
    get:
      operationId: getOrdersReport
      summary: Отчёт по заявкам
      tags: [Reports]
      description: Доступно для ролей admin и manager. Возвращает агрегированные данные по завершённым заявкам за период.
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, ready_for_dispatch, assigned, in_transit, arrived, completed, cancelled]
        - name: driverId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: warehouseId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Отчёт сформирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "403":
          description: Нет доступа
  /reports/dashboard:
    get:
      operationId: getDashboard
      summary: Дашборд менеджмента
      description: Доступно для ролей admin и manager. Возвращает агрегированную статистику по выручке, заявкам и водителям.
      tags: [Reports]
      responses:
        "200":
          description: Данные дашборда
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "403":
          description: Нет доступа